move logic and use proper variable naming convention
